+++ CHANGELOG.md
@@ -24,6 +24,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![CLI] Remove cache folder option
 - ![CLI] Replace position column with index column in `list/concepts` command
 - ![CLI] Split command `list/concepts` into `list/dimensions` and `list/attributes`
+- ![CLI] Improve startup time when using custom sources
 
 ### Fixed